Allow ordering on 'string_for_sorting.sort'
Ordering on string fields is disabled by default in es5, marking the
field as "fielddata" enables this.

The documentation warns that this can have a significant memory
penalty:

> Fielddata can consume a lot of heap space, especially when loading
> high cardinality text fields. Once fielddata has been loaded into
> the heap, it remains there for the lifetime of the segment. Also,
> loading fielddata is an expensive process which can cause users to
> experience latency hits. This is why fielddata is disabled by
> default.

...however as our schemas distinguish between sortable text and
regular text, this is probably fine.

There's also a question of whether it actually makes sense to sort on
a string field anyway:

> Before you enable fielddata, consider why you are using a text field
> for aggregations, sorting, or in a script. It usually doesn’t make
> sense to do so.
>
> A text field is analyzed before indexing so that a value like New
> York can be found by searching for new or for york. A terms
> aggregation on this field will return a new bucket and a york
> bucket, when you probably want a single bucket called New York.
>
> Instead, you should have a text field for full text searches, and an
> unanalyzed keyword field with doc_values enabled for aggregations

Possibly a better approach is to see what exactly the
"string_for_sorting" analyser is doing, and see if we can just make
this a keyword field.