Use top-level 'post_filter' on overall query
This is necessary to make aggregations work, as a 'filter' inside a
'query' is applied *before* any aggregations are computed.  What we
instead want to happen is this:

1. Restrict results by query
2. Calculate aggregations
3. Further restrict by any filters

This matches the old behaviour as described by the tests.

Strangely, switching to 'query' / 'no_match_query' (which is required)
in FormatMigrator seemed to break post-filtering.  I tracked this down
to the use of 'should'.  Turning these into 'must' fixed the problem.