Add index configuration from nodes
Previously, indices would inherit default configuration from the
nodes.  This is no longer the case.

This configuration was defined in puppet, now it's here.  This is a
bit unfortunate though in that it means we can't (easily) change the
number of replicas in the dev VM or CI (both of which were set to 0).