Implement index locking with Redis
AWS managed ES doesn't support locking, but we need it to safely
implement schema migrations.

The schema migration process locks the old index, copies data to the
new index, compares the two indices, and then swaps the alias.
Without locking, the comparison step cannot be implemented reliably.

The comparison step is essential because, eg, if ES has a problem and
rejects writes to the new index, the comparison step is what will
catch that and prevent the new index from being activated.

This change does introduce a potential TOCTOU bug though:

1. Thread 1 checks if the index is locked, and gets "no"
2. Thread 2 locks the index and begins schema migration
3. Thread 1 writes to the index
4. Thread 2 finishes schema migration, finds a discrepancy between the
   indices, and throws an error.

To offset that problem, locking imposes a 2s delay, which should make
this far less likely.