Bump limit of default sidekiq queue to 8
We've occasionally seen a build up of sidekiq jobs (in both search-api
and rummager).  I experimented by bumping the limit on the "default"
queue to 8 via the app console during one such spike: the jobs cleared
much faster than in rummager (which still had the limit of 4), and the
Elasticsearch search latency increased by maybe a couple of
milliseconds - it's hard to say because any increase is small enough
to be obscured by the natural variability of the metric.

These limits were added to solve a problem, but that problem occurred
with an almost totally different search architecture, so I think it's
worth experimenting with the limits a bit.  They could perhaps be
increased further, but let's see how this change performs for now.