Reject long queries
We're seeing a few long, slow-running, queries which are clearly not
something entered by a human.  Stemming, synonyms, and other analyses
aren't free - so spending 1.3s executing a 1677-character query which
no human is looking at the result of seems like a waste of time and
effort.

We don't want to impact human users, so the new limit is quite high,
and there will be follow-up work in finder-frontend to suggest the
contact form when someone enters a long query.