Reduce sitemap limit to 25,000 links
We've seen issues with the sitemap generation task running out of
memory.  Not sure why this has started to happen over the last few
days, but a quick fix is to reduce the number of links per file - as
the entire xml file is generated in memory before being written to
disk.