Remove conflicting test
The test for the presence of facet group links for the finder presenter
has been moved to the relevant spec for the finder facet group presenter.
This test was not present in Rummager when 7ada0beeceb06ef00f2075d7058ed63359c676d1
was merged, so tests passed. It was present in search-api so this test
failed as the implementation was moved. We still have the same test coverage
so removing this test will fix the conflict.